add 2370 prod variant
